pyproject.toml
[project]
name = "cereals-price-prediction"
version = "0.1.0"
description = ""
authors = [
    {name = "Flavian Reignault",email = "flavian.reignault@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "google-cloud-storage (>=2.19.0,<3.0.0)",
    "prophet (>=1.1.6,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
